requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pandas
pydantic